requirements.txt
Django==1.6.5
python-dateutil==2.2
requests==2.3.0